binary to decimal

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -10,7 +10,7 @@
  */
 unsigned int binary_to_uint(const char *b)
 {
-	unsigned int sum = 0, decimalVal = 1;
+	int sum = 0, decimalVal = 1;
 	int len = strlen(b), i;
 
 	if (b == NULL)
@@ -30,5 +30,5 @@ unsigned int binary_to_uint(const char *b)
 		}
 		decimalVal *= 2;
 	}
-	return (sum);
+	return ((unsigned int)sum);
 }
